LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/page-load/svg-script-src.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: */*\r\nReferer: http://web-platform.test:8800/html/syntax/speculative-parsing/gen...


FAIL Speculative parsing, page load: svg-script-src Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: */*\r\nReferer: http://web-platform.test:8800/html/syntax/speculative-parsing/generated/page-load/resources/svg-script-src-framed.sub.html?uuid=0819fd3f-6be9-4c48-abb4-1213331a3423"
